backend/requirements.txt
fastapi
uvicorn
langchain
spacy
faiss-cpu
beautifulsoup4
python-multipart
rank_bm25
langchain-core
nltk
pdfplumber
cohere
python-dotenv
psutil
networkx